feat(filter-tags) Updated structure and completed demo
