+++ apps/angular-test-app/src/app/app.component.html
@@ -17,9 +17,6 @@
         <a mat-list-item routerLink="/popover" routerLinkActive="uxg-list-item-selected">
           <p>Popover</p>
         </a>
-        <a mat-list-item routerLink="/filter-tags" routerLinkActive="uxg-list-item-selected">
-          <p>Filter Tags</p>
-        </a>
         <a mat-list-item routerLink="/filter-panel" routerLinkActive="uxg-list-item-selected">
           <p>Filter Panel</p>
         </a>
