+++ apps/angular-test-app/src/app/components/filter-panel-demo/filter-panel-demo.component.scss
@@ -11,7 +11,7 @@
 .filter-panel-side {
     width: auto !important;
     min-width: 352px;
-    max-width: 450px;
+    max-width: 352px;
 }
 
 .filter-panel-side-header {
@@ -56,7 +56,7 @@
 
 .filter-panel-tags {
     padding: 0 25px 0 25px;
-
+    
     mat-form-field {
         width: 100%;
     }
