--- apps/angular-test-app/src/app/components/filter-tags-demo/filter-tags-demo.component.html
@@ -1,7 +1,0 @@
-<div class="ffdc-filter-tags-demo-div">
-  <!-- <div class="ffdc-filter-tags-demo-title">Filter Tags Demo</div> -->
-
-  <uxg-filter-tags [data]="dataDemo" (change)="changeFilter($event)"> </uxg-filter-tags>
-
-  <pre>{{ chosenTags | json }}</pre>
-</div>
