--- apps/angular-test-app/src/app/components/filter-tags-demo/filter-tags-demo.component.scss
@@ -1,7 +1,0 @@
-.ffdc-filter-tags-demo-div {
-    width: 352px;
-    height: 196px; 
-}
-.ffdc-filter-tags-demo-title {
-    margin: 10px 0;
-}
