+++ apps/angular-test-app/src/app/components/home/home.component.html
@@ -57,17 +57,6 @@ imports: [
           </mat-card-subtitle>
         </mat-card-title-group>
       </mat-card>
-      <mat-card routerLink="/filter-tags">
-        <img mat-card-image src="assets/bar-chart.png" alt="Abstract illustration of an avatar" />
-        <mat-card-title-group>
-          <mat-card-title class="uxg-h4">Filter Tags</mat-card-title>
-          <mat-card-subtitle class="uxg-overline">
-            <mat-chip-list>
-              <mat-chip disabled dense>FilterTags</mat-chip>
-            </mat-chip-list>
-          </mat-card-subtitle>
-        </mat-card-title-group>
-      </mat-card>
       <mat-card routerLink="/filter-panel">
         <img mat-card-image src="assets/bar-chart.png" alt="Abstract illustration of an avatar" />
         <mat-card-title-group>
