--- libs/angular-components/filter-tags/ng-package.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/libs/angular-components/filter-tags",
-  "lib": {
-    "entryFile": "index.ts"
-  }
-}
