--- libs/angular-components/filter-tags/src/_filter-tags.theme.scss
@@ -1,108 +1,0 @@
-@mixin uxg-filter-tags-theme($theme) {
-  $primary: map-get($theme, primary);
-  $foreground: map-get($theme, foreground);
-  $background: map-get($theme, background);
-
-  $filter-tags-primary-color: mat-color($primary, default);
-  $filter-tags-text-color: map-get($foreground, text);
-  $filter-tags-background-color: mat-color($primary, default, .15);
-  $filter-tags-background-color-remove-icon: mat-color($primary, default, .50);
-  $filter-tags-text-color-clear: mat-color($uxg-fuchsia, 500);
-
-  .input-container {
-    width: 100%;
-    box-sizing: border-box;
-    margin: 2px 0;
-    border: 2px solid $filter-tags-primary-color;
-    border-radius: 4px;
-  }
-
-  .mat-icon.search-icon {
-    color: $filter-tags-primary-color;
-    margin: 16px 5px 14px 17px;
-    font-size: 17.49px !important;
-    line-height: 18px !important;
-    width: 18px !important;
-    height: 18px !important;
-  }
-
-  .mat-icon {
-    width: 14px !important;
-    height: 14px !important;
-    font-size: 14px !important;
-    line-height: 14px !important;
-  }
-
-  .mat-chip {
-    color: $filter-tags-primary-color;
-    font-size: 12px;
-    font-weight: 300;
-    line-height: 18px;
-    font-family: Roboto, sans-serif;
-  }
-
-  input.mat-input-element {
-    width: 80%;
-    margin: 13px 0px 11px 5px;
-    font-size: 16px;
-    line-height: 24px;
-    letter-spacing: 0.16px;
-    font-family: Roboto, sans-serif;
-  }
-
-
-  .mat-chip.mat-standard-chip {
-    background-color: $filter-tags-background-color;
-    border: 1px solid $filter-tags-primary-color;
-    border-radius: 18px;
-    height: 24px;
-    padding: 3px 8px 3px 12px;
-    margin: 0px 4px 16px 4px;
-  }
-
-  .mat-chip.mat-standard-chip .mat-chip-remove {
-    color: $filter-tags-background-color-remove-icon;
-    margin: 5px 0px 5px 8px;
-  }
-
-  .mat-chip.mat-standard-chip.mat-chip-selected.mat-accent .mat-chip-remove {
-    opacity: 1;
-  }
-
-  .mat-chip.mat-standard-chip.mat-chip-selected.mat-accent {
-    background-color: $filter-tags-background-color;
-    color: $filter-tags-primary-color;
-  }
-
-  .clear-filters {
-    color: $filter-tags-text-color-clear;
-    font-size: 12px;
-    font-weight: 300;
-    line-height: 18px;
-    border-width: 0px;
-    background-color: transparent;
-    float: right;
-  }
-
-  .filter-tags-title {
-    color: $filter-tags-text-color;
-    display: inline;
-  }
-
-  .filter-tags-title h5 {
-    display: inherit;
-  }
-
-  .mat-option {
-    color: $filter-tags-text-color;
-    font-size: 14px;
-    font-weight: 300;
-    line-height: 20px;
-  }
-
-  .selected-option-highlight,
-  .selected-option-highlight.mat-selected {
-    background-color: $filter-tags-background-color !important;
-    font-weight: 400;
-  }
-}
