--- libs/angular-components/filter-tags/src/filter-tags.component.html
@@ -1,52 +1,0 @@
-<div class="filter-tags-block">
-  <div class="filter-tags-header">
-    <div class="filter-tags-title"><h5 class="uxg-h5">Tags</h5></div>
-    <button class="clear-filters" (click)="clearSelection()">
-      Clear all
-    </button>
-  </div>
-  <div class="filter-tags-body">
-    <mat-form-field class="chip-list" appearance="standard">
-      <mat-chip-list #chipList aria-label="Tags selection">
-        <mat-chip dense selected *ngFor="let tag of selectedData" [removable]="removable" (removed)="remove(tag)">
-          {{ tag }}
-          <mat-icon matChipRemove *ngIf="removable">cancel</mat-icon>
-        </mat-chip>
-      </mat-chip-list>
-
-      <div class="input-container">
-        <mat-icon matPrefix class="search-icon">search</mat-icon>
-        <input
-          matInput
-          type="text"
-          placeholder="{{ inputPlaceholder }}"
-          #tagInput
-          [formControl]="tagCtrl"
-          [matAutocomplete]="auto"
-          [matChipInputFor]="chipList"
-          [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
-          [matChipInputAddOnBlur]="addOnBlur"
-          (matChipInputTokenEnd)="add($event)"
-          (click)="onClick()"
-        />
-      </div>
-      <mat-chip-trailing-icon></mat-chip-trailing-icon>
-      <mat-autocomplete
-        #auto="matAutocomplete"
-        (optionSelected)="selected($event)"
-        (opened)="onOpen($event)"
-        (closed)="onClose($event)"
-      >
-        <mat-option
-          *ngFor="let tag of (filteredTags | async)"
-          [value]="tag"
-          [ngClass]="{
-            'selected-option-highlight': selectedData.indexOf(tag.trim()) > -1
-          }"
-        >
-          {{ tag }}
-        </mat-option>
-      </mat-autocomplete>
-    </mat-form-field>
-  </div>
-</div>
