+++ libs/angular-components/filter/filter-tags/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/libs/angular-components/filter-panel",
+  "lib": {
+    "entryFile": "index.ts"
+  }
+}
