+++ libs/angular-components/filter/filter-tags/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ffdc/uxg-angular-components/filter-tags",
+  "name": "@ffdc/uxg-angular-components/filter/filter-panel",
   "version": "0.13.4",
   "author": "u721212 <adrian.apetrei.89@gmail.com>"
 }
