+++ libs/angular-components/filter/filter-tags/src/filter-tags.component.html
@@ -1,0 +1,36 @@
+<div class="filter-tags-container">
+  <mat-chip-list #chipList aria-label="Tags selection">
+    <mat-chip dense selected *ngFor="let tag of selectedData" [removable]="removable" (removed)="remove(tag)">
+      {{ tag }}
+      <mat-icon matChipRemove *ngIf="removable">cancel</mat-icon>
+    </mat-chip>
+  </mat-chip-list>
+
+  <mat-form-field class="chip-list" appearance="outline">
+    <mat-label><mat-icon>search</mat-icon> Filter by tags</mat-label>
+    <input
+      matInput
+      type="text"
+      placeholder="{{ placeholder }}"
+      #tagInput
+      [formControl]="formCtrl"
+      [matAutocomplete]="auto"
+      [matChipInputFor]="chipList"
+      [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
+      [matChipInputAddOnBlur]="addOnBlur"
+      (matChipInputTokenEnd)="add($event)"
+      (click)="onClick()"
+    />
+  </mat-form-field>
+  <mat-autocomplete #auto="matAutocomplete" (optionSelected)="selected($event)">
+    <mat-option
+      *ngFor="let tag of (filteredTags$ | async)"
+      [value]="tag"
+      [ngClass]="{
+        'selected-option-highlight': selectedData.indexOf(tag.trim()) > -1
+      }"
+    >
+      {{ tag }}
+    </mat-option>
+  </mat-autocomplete>
+</div>
