+++ libs/angular-components/filter/filter-tags/src/filter-tags.component.scss
@@ -1,0 +1,8 @@
+@import '../../../../../themes/angular-theme/uxg-app-theme';
+@import './filter-tags.theme';
+@include uxg-filter-tags-theme($uxg-light-theme);
+
+
+
+
+
