+++ libs/angular-components/filter/filter-tags/tsconfig.lib.json
