+++ libs/angular-components/filter/filter-tags/tslint.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../tslint.json",
+  "extends": "../../tslint.json",
   "rules": {
     "directive-selector": [true, "attribute", "uxg", "camelCase"],
     "component-selector": [true, "element", "uxg", "kebab-case"]
