+++ libs/angular-components/index.ts
@@ -3,4 +3,4 @@ export * from './popover';
 export * from './scroll-to-top';
 export * from './table';
 export * from './filter/filter-tree';
-export * from './filter-tags';
+export * from './filter/filter-tags';
